Fix TM2 to startup when cache bytes don't change
This fixes an issue where TM refused to start if a cache's bytes in
astats never change. TM returns 503 until it's polled every cache
twice, and has a BytesPerSecond calculation. The issue was, if a cache
had hard-coded proc.net.dev which never changed, TM has no way to know
the cache simply hasn't updated its stats, and that we won't get a
different stat on the next poll.

This gives TM a timeout, once it successfully polls a cache once,
if the stats don't change in 60s, we start up anyway. It's not
reasonable for a cache to update stats less frequently than a minute.
While guessing isn't ideal, there's no good alternative. We can't
know the cache simply hasn't updated its stats internally; the
only option is a timeout, or never starting.

Incidentally, this problem occurred in production. We had a bad
cache lying to us with hard-coded astats that never updated, causing
TM to never start.